Refactor: Improve code quality and enforce stricter static analysis
- Strengthened compiler warnings and integrated clang-tidy into the
  build
- Improved const correctness and exception safety guarantees
- Replaced low-level POSIX patterns with project abstractions in network
  code
- Eliminated compiler warnings across all translation units
- Removed dead includes and reduced unnecessary header dependencies

+++ include/string_util.hpp
@@ -35,14 +35,14 @@ namespace StringUtil {
         // Single definition to avoid repeated instantiation in each template
         inline constexpr std::array<char, 256> lower_table = []() noexcept {
             std::array<char, 256> table{};
-            for (int i = 0; i < 256; ++i)
+            for (size_t i = 0; i < 256; ++i)
                 table[i] = (i >= 'A' && i <= 'Z') ? static_cast<char>(i + 32) : static_cast<char>(i);
             return table;
         }();
 
         inline constexpr std::array<char, 256> upper_table = []() noexcept {
             std::array<char, 256> table{};
-            for (int i = 0; i < 256; ++i)
+            for (size_t i = 0; i < 256; ++i)
                 table[i] = (i >= 'a' && i <= 'z') ? static_cast<char>(i - 32) : static_cast<char>(i);
             return table;
         }();
