+++ src/dns/util.hpp
@@ -19,7 +19,8 @@ namespace DNS::Util {
 
     /// Read a 16-bit big-endian value from a raw pointer.
     [[nodiscard]] constexpr std::uint16_t read_u16_be(const std::uint8_t *buf) noexcept {
-        return (static_cast<std::uint16_t>(buf[0]) << 8) | buf[1];
+        return static_cast<std::uint16_t>(
+            (static_cast<std::uint16_t>(buf[0]) << 8) | static_cast<std::uint16_t>(buf[1]));
     }
 
     /// Read a 16-bit big-endian value from a span at the given offset.
